+++ code/Rv32T/LEDR.h
@@ -7,7 +7,7 @@ class LEDR : public Device
 	T_RTTI_CLASS;
 
 public:
-	virtual void eval(VSoC* soc, uint32_t time) override final;
+	virtual void eval(VSoC* soc, uint64_t time) override final;
 
 private:
 	uint32_t m_current = 0;
