+++ code/Rv32T/TracePC.h
@@ -9,7 +9,7 @@ class TracePC : public Device
 public:
 	explicit TracePC(traktor::OutputStream* os);
 
-	virtual void eval(VSoC* soc, uint32_t time) override final;
+	virtual void eval(VSoC* soc, uint64_t time) override final;
 
 private:
 	traktor::Ref< traktor::OutputStream > m_os;
