+++ code/Rv32T/UART_TX.h
@@ -7,7 +7,7 @@ class UART_TX : public Device
 	T_RTTI_CLASS;
 
 public:
-	virtual void eval(VSoC* soc, uint32_t time) override final;
+	virtual void eval(VSoC* soc, uint64_t time) override final;
 
 private:
 	int32_t m_state = 0;
